PUSH-H-0.3.1: textures is fixed next step fix norm

+++ cub3D.h
@@ -6,7 +6,7 @@
 /*   By: hibouzid <hibouzid@student.1337.ma>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/31 21:17:06 by hibouzid          #+#    #+#             */
-/*   Updated: 2024/09/06 18:10:46 by hibouzid         ###   ########.fr       */
+/*   Updated: 2024/09/07 17:17:53 by hibouzid         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -134,7 +134,7 @@ int				mlx_get_color(t_data *data, int x, int y);
 int				valid_textures(t_data *data);
 void			ft_free_map(t_data	*data);
 int				valid_color_pre(char *p, int is_color);
-int raycasting(t_data *data, float ang, int *v_f);
+float raycasting(t_data *data, float ang, int *v_f);
 int	check_vr_hr(t_data *data, float x, float y, float Angle);
 
 #endif
